+++ cogs/event_handler.py
@@ -1,5 +1,4 @@
 import datetime
-from difflib import SequenceMatcher
 
 import basic_logger as bl
 import nextcord
@@ -169,33 +168,6 @@ class event_handler(commands.Cog, name="Event Handler"):
             await dm.send(embed=embed)
         except nextcord.errors.HTTPException:
             pass
-
-    @commands.Cog.listener()
-    async def on_command_error(self, ctx: Context, error: commands.CommandError):
-        """Called when a command error occurs"""
-        if isinstance(error, commands.CommandNotFound):
-            message = ctx.message.content[1:].split(" ")[0]
-            cmds = {
-                command: round(SequenceMatcher(None, command, message).ratio() * 100, 1)
-                for command in self.bot.all_commands.keys()
-            }
-            sorted_keys = list(sorted(cmds, key=cmds.get))
-            sorted_keys.reverse()
-            sorted_cmds = {w: cmds[w] for w in sorted_keys}
-            best_cmd = list(sorted_cmds.keys())[0]
-            best_cmd_2 = ""
-            best_cmd_full = self.bot.get_command(best_cmd)
-            if best_cmd != best_cmd_full.name and best_cmd_full is not None:
-                best_cmd_2 = f" ({best_cmd_full})"
-            best_cmd_perc = list(sorted_cmds.values())[0]
-            error = (
-                str(error)
-                + f"\nDid you mean `.{best_cmd}{best_cmd_2}`? ({best_cmd_perc}% match)"
-            )
-        embed = nextcord.Embed(color=0xFF0000)
-        embed.add_field(name="An error occured!", value=error, inline=True)
-        await ctx.send(embed=embed)
-        bl.error(error, __file__)
 
 
 def setup(bot: commands.Bot):
